major improvements in hosted service, in UserClient.cs, auto reg works now
