+++ RPC.Network/PacketProcessor.cs
@@ -5,12 +5,12 @@ using RPC.Contracts.Interfaces;
 
 namespace RPC.Network;
 
-public class PacketProcessor : IPacketProcessor
+public class PacketProcessor(IOperationRegistry operationRegistry) : IPacketProcessor
 {
     private readonly ConcurrentDictionary<Type, Func<UserClient, object, Task>> _handlers = new();
-    private readonly ConcurrentDictionary<int, Type> _opIdToPacketType = new();
-
-
+    private readonly IOperationRegistry _operationRegistry = operationRegistry
+                                                             ?? throw new ArgumentNullException(nameof(operationRegistry));
+    
     public void RegisterHandler<TPacket>(Func<UserClient, TPacket, Task> handler) where TPacket : class
     {
         _handlers[typeof(TPacket)] = async (client, packet) =>
@@ -21,65 +21,41 @@ public class PacketProcessor : IPacketProcessor
             }
         };
     }
-
-
+    
     public void RegisterHandler(Type packetType, int operationId, Func<UserClient, object, Task> handler)
-    {
-        _opIdToPacketType[operationId] = packetType;
-        _handlers[packetType] = handler;
-    }
-
+        => _handlers[packetType] = handler
+                                   ?? throw new ArgumentNullException(nameof(handler));
 
     public async Task ProcessPacketAsync<TPacket>(UserClient client, TPacket packet) where TPacket : class
     {
-        try
+        if (_handlers.TryGetValue(typeof(TPacket), out var handler))
         {
-            if (_handlers.TryGetValue(typeof(TPacket), out var handler))
-            {
-                await handler(client, packet).ConfigureAwait(false);
-            }
-            else
-            {
-                throw new Exception($"No handler registered for {typeof(TPacket)}");
-            }
+            await handler(client, packet).ConfigureAwait(false);
+            return;
         }
-        catch
-        {
-            //debug
-            throw;
-        }
+        
+        throw new InvalidOperationException($"No handler registered for {typeof(TPacket)}");
     }
-
+    
     public async Task ProcessRawPacketAsync(UserClient client, int operationId, Guid requestId, string packetData)
     {
-        try
+        if (!_operationRegistry.TryGetPacketType(operationId, out var packetType) || packetType == null)
         {
-            if (!_opIdToPacketType.TryGetValue(operationId, out var packetType) || packetType == null)
-            {
-                throw new Exception($"No handler registered for operation id {operationId}");
-            }
-
-
-            var packet = JsonSerializer.Deserialize(packetData, packetType);
-            if (packet == null)
-            {
-                throw new Exception($"Failed to deserialize packet for operation id {operationId}");
-            }
-
-
-            if (_handlers.TryGetValue(packetType, out var handler))
-            {
-                await handler(client, packet).ConfigureAwait(false);
-            }
-            else
-            {
-                throw new Exception($"No handler registered for packet type {packetType}");
-            }
+            throw new InvalidOperationException($"No handler registered for operation id {operationId}");
         }
-        catch
+
+        var packet = JsonSerializer.Deserialize(packetData, packetType);
+        if (packet == null)
         {
-            //debug
-            throw;
+            throw new InvalidDataException($"Failed to deserialize packet for operation id {operationId}");
         }
+
+        if (_handlers.TryGetValue(packetType, out var handler))
+        {
+            await handler(client, packet).ConfigureAwait(false);
+            return;
+        }
+        
+        throw new InvalidOperationException($"No handler registered for packet type {packetType}");
     }
 }
