+++ RpcApi/appsettings.json
@@ -1,6 +1,6 @@
 {
   "RpcSettings" : {
-    "Host": "",
+    "Host": "0.0.0.0",
     "Port": 5005
   },
   "Logging": {
